+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: ff5a848 - v147
+| Author: fjrodafo - Sun, 19 Jan 2025 23:29:38 +0000
+| 
 * Commit: 58cb4bf - v146
 | Author: fjrodafo - Sat, 18 Jan 2025 22:05:45 +0000
 |